requirements.txt
bcrypt
fastapi
pydantic
python-jose
PyJWT
uvicorn
passlib
sqlalchemy
pymysql